+++ LookupAnything/docs/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## Upcoming release
+* Add upper-right close buttons for lookup and search menus.
 * Improved translations. Thanks to BlackRosePetals (updated Chinese) and creeperkatze (updated German)!
 
 ## 1.48.0